Fix input reading; alternative to GH-5
In case an input file does not have a new line character on its last line, the last f.readline() call does not have this newline.
Therefore, line[:-1] will remove the character ']' that should indicate that the lattice parsed completely, i.e., the format is [[v1] [v2] ... [vn]].
This caused the code to crash: another line was expected because it did not detect the end of input "]]".
By switching line[:-1] to line.strip(), we support reading a line without a newline at its end.